// src/pages/visa-calculation/components/calculation-card/types.d.ts
export type FormValue = {
  test_visa: string
  age: {
    point: number
    value: strings
  }
  proficiency: {
    point: number
    value: strings
  }
  language: string
  additional_information: {
    point: number
    value: strings
  }
  experience_in_korea: {
    point: number
    value: strings
  }
  research: {
    point: number
    value: strings
  }
  exchange: {
    point: number
    value: strings
  }
  training: {
    point: number
    value: strings
  }
  topik_or_kiip: {
    point: number
    value: strings
  }
  private_training: {
    point: number
    value: strings
  }
  language_training: {
    point: number
    value: strings
  }
  additional_points: {
    point: number
    value: strings
  }
  immigration: {
    point: number
    value: strings
  }
  punishment: {
    point: number
    value: strings
  }
}